reorganize tcgc staff
