+++ packages/typespec-client-generator-core/src/internal-utils.ts
@@ -40,7 +40,7 @@ import {
   getVersioningMutators,
   getVersions,
 } from "@typespec/versioning";
-import { getParamAlias, listOperationGroups } from "./decorators.js";
+import { getParamAlias } from "./decorators.js";
 import {
   DecoratorInfo,
   SdkBuiltInType,
@@ -48,7 +48,6 @@ import {
   SdkEnumType,
   SdkHttpResponse,
   SdkModelPropertyType,
-  SdkOperationGroup,
   SdkType,
   TCGCContext,
 } from "./interfaces.js";
@@ -648,22 +647,6 @@ export function handleVersioningMutationForGlobalNamespace(context: TCGCContext)
   const subgraph = unsafe_mutateSubgraphWithNamespace(context.program, [mutator], globalNamespace);
   compilerAssert(subgraph.type.kind === "Namespace", "Should not have mutated to another type");
   return subgraph.type;
-}
-
-export function listRawSubClients(
-  context: TCGCContext,
-  client: SdkOperationGroup | SdkClient,
-): SdkOperationGroup[] {
-  const retval: SdkOperationGroup[] = [];
-  const queue: SdkOperationGroup[] = listOperationGroups(context, client);
-  while (queue.length > 0) {
-    const operationGroup = queue.pop()!;
-    retval.push(operationGroup);
-    if (operationGroup.subOperationGroups) {
-      queue.push(...operationGroup.subOperationGroups);
-    }
-  }
-  return retval;
 }
 
 export function resolveDuplicateGenearatedName(
