+++ packages/typespec-client-generator-core/src/types.ts
@@ -54,6 +54,7 @@ import {
   getOverriddenClientMethod,
   getUsageOverride,
   listClients,
+  listOperationGroups,
   listOperationsInOperationGroup,
   shouldFlattenProperty,
   shouldGenerateConvenient,
@@ -104,7 +105,6 @@ import {
   isNeverOrVoidType,
   isOnClient,
   listAllUserDefinedNamespaces,
-  listRawSubClients,
   resolveConflictGeneratedName,
   twoParamsEquivalent,
   updateWithApiVersionInformation,
@@ -1911,7 +1911,7 @@ export function handleAllTypes(context: TCGCContext): [void, readonly Diagnostic
       // operations on a client
       diagnostics.pipe(updateTypesFromOperation(context, operation));
     }
-    for (const sc of listRawSubClients(context, client)) {
+    for (const sc of listOperationGroups(context, client, true)) {
       for (const operation of listOperationsInOperationGroup(context, sc)) {
         // operations on operation groups
         diagnostics.pipe(updateTypesFromOperation(context, operation));
