+++ packages/typespec-client-generator-core/src/validations/clients.ts
@@ -1,0 +1,3 @@
+import { TCGCContext } from "../interfaces.js";
+
+export function validateClients(context: TCGCContext) {}
