+++ packages/typespec-client-generator-core/src/validations/http.ts
@@ -1,0 +1,3 @@
+import { TCGCContext } from "../interfaces.js";
+
+export function validateHttp(context: TCGCContext) {}
