+++ packages/typespec-client-generator-core/src/validations/methods.ts
@@ -1,0 +1,3 @@
+import { TCGCContext } from "../interfaces.js";
+
+export function validateMethods(context: TCGCContext) {}
