+++ packages/typespec-client-generator-core/test/decorators/client.test.ts
@@ -732,9 +732,9 @@ describe("listOperationGroups without @client and @operationGroup", () => {
     deepStrictEqual(listOperationGroups(runner.context, ag), []);
 
     allOperationGroups = listOperationGroups(runner.context, client, true);
-    deepStrictEqual(allOperationGroups, [aa, aaa, aab, aabGroup1, aabGroup2, aag, ag]);
+    deepStrictEqual(allOperationGroups, [aa, ag, aaa, aab, aag, aabGroup1, aabGroup2]);
     allOperationGroups = listOperationGroups(runner.context, aa, true);
-    deepStrictEqual(allOperationGroups, [aaa, aab, aabGroup1, aabGroup2, aag]);
+    deepStrictEqual(allOperationGroups, [aaa, aab, aag, aabGroup1, aabGroup2]);
     allOperationGroups = listOperationGroups(runner.context, aaa, true);
     deepStrictEqual(allOperationGroups, []);
     allOperationGroups = listOperationGroups(runner.context, aab, true);
@@ -865,38 +865,6 @@ describe("listOperationGroups without @client and @operationGroup", () => {
 });
 
 describe("client hierarchy", () => {
-  it("multi clients ", async () => {
-    await runner.compile(`
-        @service
-        namespace Test1Client {
-          op x(): void;
-        }
-        @service
-        namespace Test2Client {
-          op y(): void;
-        }
-      `);
-
-    const clients = listClients(runner.context);
-    deepStrictEqual(clients.length, 2);
-
-    const client1 = clients.find((x) => x.name === "Test1Client");
-    ok(client1);
-    deepStrictEqual(
-      listOperationsInOperationGroup(runner.context, client1).map((x) => x.name),
-      ["x"],
-    );
-    deepStrictEqual(listOperationGroups(runner.context, client1).length, 0);
-
-    const client2 = clients.find((x) => x.name === "Test2Client");
-    ok(client2);
-    deepStrictEqual(
-      listOperationsInOperationGroup(runner.context, client2).map((x) => x.name),
-      ["y"],
-    );
-    deepStrictEqual(listOperationGroups(runner.context, client2).length, 0);
-  });
-
   it("no client", async () => {
     await runner.compile(`
         namespace Test1Client {
