+++ packages/typespec-client-generator-core/test/package/azure-widget-service.test.ts
