+++ packages/typespec-client-generator-core/test/package/models-only.test.ts
