+++ packages/typespec-client-generator-core/test/package/vanilla-widget-service.test.ts
