+++ website/src/content/docs/docs/howtos/Generate client libraries/07tcgcTypes.mdx
@@ -51,278 +51,6 @@ interface Widgets {
 }
 ```
 
-## Overall Hierarchy of Types
-
-This is a rough sketch of the hierarchy. We'll go into more detail for each section, this is just to get an idea of the overall structure.
-
-```ts
-const sdkContext = {
-  sdkPackage: {
-    clients: [
-      {
-        kind: "client",
-        name: "WidgetClient",
-        initialization: {
-          properties: [
-            {
-              kind: "endpoint",
-              type: {
-                kind: "union",
-                values: [
-                  {
-                    kind: "endpoint",
-                    serverUrl: "{endpoint}",
-                    templateArguments: [
-                      {
-                        name: "endpoint",
-                        kind: "path",
-                        ...
-                      }
-                    ]
-                  },
-                  {
-                    kind: "endpoint",
-                    serverUrl: "{endpoint}/widget",
-                    templateArguments: [
-                      {
-                        name: "endpoint",
-                        kind: "path",
-                        ...
-                      }
-                    ]
-                  }
-                ]
-              }
-            },
-            {
-              kind: "credential",
-              ...
-            }
-          ]
-        },
-        methods: [
-          {
-            kind: "clientaccessor",
-            response: {
-              name: "WidgetPartsClient",
-              kind: "client",
-              ...
-            }
-          },
-          {
-            kind: "basic",
-            parameters: [
-              {
-                name: "id",
-                kind: "method",
-                type: {
-                  kind: "string",
-                },
-                ...
-              }
-            ],
-            operation: {
-              kind: "http",
-              verb: "get",
-              parameters: [
-                {
-                  name: "id",
-                  kind: "path",
-                  type: {
-                    kind: "string",
-                  },
-                  correspondingMethodParams: [
-                    {
-                      name: "id",
-                      kind: "method",
-                      type: {
-                        kind: "string",
-                      },
-                      ...
-                    }
-                  ]
-                }
-              ],
-              bodyParam: undefined,
-              responses: {
-                200: {
-                  kind: "http",
-                  ...
-                }
-              },
-              examples: [
-                {
-                  kind: "http",
-                  name: "Example 1",
-                  description: "Example 1",
-                  filePath: "example1.json",
-                  rawExample: {...},
-                  parameters: [...],
-                  responses: {...}
-                }
-              ]
-            },
-            response: {
-              kind: "method",
-              logicalPath: undefined,
-              ...
-            },
-            exception: {
-              kind: "method",
-              ...
-            },
-            ...
-          }
-        ],
-        ...
-      }
-    ],
-    models: [
-      {
-        kind: "model",
-        name: "Widget",
-        properties: [
-          {
-            kind: "path",
-            name: "id",
-            type: {
-              kind: "string",
-            },
-            ...
-          },
-          {
-            kind: "property",
-            name: "color",
-            type: {
-              kind: "enum",
-              isFixed: true,
-              ...
-            },
-            ...
-          }
-        ]
-      }
-    ],
-    enums: [
-      {
-        name: "WidgetColor",
-        generatedName: true,
-        values: [
-          {
-            value: "red",
-            ...
-          },
-          {
-            value: "blue",
-            ...
-          }
-        ]
-      }
-    ],
-    ...
-  },
-  ...
-}
-```
-
-## SdkContext
-
-`SdkContext` is the context of your SDK emitter.
-
-### Interface
-
-```ts
-export interface SdkContext<
-  TOptions extends object = Record<string, any>,
-  TServiceOperation extends SdkServiceOperation = SdkHttpOperation,
-> {
-  // emit context from tsp that is used to create our TCGC Sdk Context
-  emitContext: EmitContext<TOptions>;
-  // representation of the entire package the language emitter should generate
-  sdkPackage: SdkPackage<TServiceOperation>;
-  program: Program;
-  emitterName: string;
-}
-```
-
-### Usage
-
-It is used to persist information across functions. You should never define global constants or variables in your emitter code, because this leads to issues with multiple emitter runs. Instead, all global information should be stored in your context.
-
-You should define your own language-specific `SdkContext` that extends from the one defined in `@azure-tools/typespec-client-generator-core`. This is where you keep your global variables. You should utilize the `createSdkContext` from tcgc.
-You also shouldn't instantiate your own interface, instead you should call `createSdkContext` to create.
-
-```ts
-import { EmitContext } from "@typespec/compiler";
-import { createSdkContext, SdkServiceOperation } from "@azure-tools/typespec-client-generator-core";
-import { PythonEmitterOptions } from "./lib.js";
-
-interface PythonSdkContext<TServiceOperation extends SdkServiceOperation>
-  extends SdkContext<PythonEmitterOptions, TServiceOperation> {
-  ClientMap: Map<SdkClientType, PythonSdkClientType>;
-}
-
-export async function $onEmit(context: EmitContext<PythonEmitterOptions>) {
-  const sdkContext: PythonSdkContext = {
-    ...createSdkContext(context, "@azure-tools/typespec-python"),
-    ClientMap: new Map<sdkClientType, PythonSdkClientType>(),
-  };
-}
-```
-
-## SdkPackage
-
-`SdkPackage` represents the entire package that your language emitter should generate. It synthesizes all of the information language emitters will need to generate the entire package
-
-### Interface
-
-```ts
-export interface SdkPackage<TServiceOperation extends SdkServiceOperation> {
-  name: string;
-  rootNamespace: string;
-  clients: SdkClientType<TServiceOperation>[];
-  models: SdkModelType[];
-  enums: SdkEnumType[];
-  diagnostics: readonly Diagnostic[];
-}
-```
-
-### Example
-
-```ts
-const sdkPackage = {
-  name: "ContosoWidgetManager",
-  rootNamespace: "Contoso.WidgetManager",
-  clients: [
-    {
-      name: "ContosoWidgetManagerClient",
-      kind: "client",
-      ...
-    }
-  ],
-  models: [
-    {
-      name: "Widget",
-      kind: "model",
-      ...
-    },
-    {
-      name: "Error",
-      kind: "model",
-      ...
-    }
-  ],
-  enums: [],
-  diagnostics: [],
-}
-```
-
-### Usage
-
-```ts
-const sdkPackage = sdkContext.sdkPackage;
-```
-
 ## SdkClientType
 
 An `SdkClientType` represents a single client in your package.
